requirements.txt
flask
openai
python-dotenv
slack-sdk
